requirements.txt
Flask==1.1.2
python-dotenv==0.14.0
gunicorn==20.0.4
gevent==20.6.2
psycopg2==2.8.5
PyJWT==1.7.1